Update money and legends SVG icons, remove redundant elements
The SVG icons for 'money' and 'legends' were modified, specifically by removing redundant g-tags and clip-path. This simplifies the SVG structure for better load performance and easier maintenance. A revision of viewBox parameters was also performed to standardize both icons dimensions. The newline character at the end of 'money' SVG was removed as part of code neat cleanup.